requirements.txt
langchain
langchain-core
langchain-community
langgraph
python-dotenv
langchain-groq
arxiv
wikipedia
langchain-tavily
langgraph-cli[inmem]
langchain-ollama
faiss-cpu
djangorestframework
markdown
django-filter
psycopg2-binary
chromadb
langchain-chroma
ipython
django
djangorestframework